requirements.txt
Flask
gunicorn
twilio
mysql-connector-python
python-dotenv